I added all the courses and made sure semester plan worked

+++ src/DataLoader.java
@@ -130,7 +130,7 @@ public class DataLoader extends DataConstants {
 				String courseName = (String) courseJSON.get(COURSE_NAME);
 				String courseDescription = (String) courseJSON.get(COURSE_DESCRIPTION);
 				boolean courseAvailability = (boolean) courseJSON.get(COURSE_AVAILABILITY);
-				double courseCredits = ((double) courseJSON.get(COURSE_CREDITS));
+				double courseCredits = ((double)(long) courseJSON.get(COURSE_CREDITS));
 				ArrayList<String> term = courseJSON.get(COURSE_TERM) != null ? (ArrayList<String>) courseJSON.get(COURSE_TERM) : new ArrayList<String>();
 				String electiveArea = courseJSON.get(COURSE_ELECTIVE_AREA) != null ? (String)courseJSON.get(COURSE_ELECTIVE_AREA) : null;
 				String passingGrade = (courseJSON.get(COURSE_PASSING_GRADE) != null) ? passingGrade = (String)(courseJSON.get(COURSE_PASSING_GRADE)) : "";
